test: tent is long

+++ src/base_lib.h
@@ -30,7 +30,7 @@ void get_arg(const char *coded, unsigned long *start, unsigned long *inc, unsign
 
 void get_current_time(struct timespec *curr_t);
 
-static inline void timespec_div(struct timespec *rop, const struct timespec *op, unsigned int div)
+static inline void timespec_div(struct timespec *rop, const struct timespec *op, unsigned long div)
 {
 	rop->tv_nsec =
 	    ((op->tv_sec % div) * TIME_PRECISION / div) + (op->tv_nsec / div);
